Removed a lot of gcc warnings.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@1793 01821946-851b-0410-b7f4-f1477c70e57a

+++ effects/lightning.cpp
@@ -463,7 +463,7 @@ if (pl && pl->pNodes) {
 
 	for (i = abs (pl->nNodes), pln = pl->pNodes; i > 0; i--, pln++) {
 		if (pln->pChild) {
-			if ((int) (size_t) pln->pChild == 0xffffffff)
+			if ((int) (size_t) pln->pChild == (int) 0xffffffff)
 				pln->pChild = NULL;
 			else {
 				DestroyLightningNodes (pln->pChild);
